Activity analysis and aesthetic improvements
Modified calculation of activity:
1. Bug fix: in previous versions, the "one minute" of activity was not
actually one minute - calculation was for stim to stim+1/bin_width,
which would give you either 1.5 minutes or 2 minutes depending on the
width of the bin. So that's pretty problematic, especially since the
width of activity to consider depends on bin size. In any case, that's
been fixed in this version.
2. Feature enhancement: activity now includes assessment of the first,
second, and third minutes after stim onset.
3. Aesthetic improvements: responsiveness plots now look a bit nicer, as
do the latency plots.
4. This version also contains a new function: plotArousalProbabilities.
Basically it gives you the ability to plot probability of arousal (for a
given fly) as a function of average prior wake duration, prior sleep
duration, etc.